# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Public read-only API for browsing Dwellings property listings.",
        "title": "Dwellings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.dwellings.tv",
    "basePath": "/",
    "paths": {
        "/api/v1/properties": {
            "get": {
                "description": "Browse active listings with filters, sorting, and cursor pagination. Pass the returned next_cursor to fetch the following page.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "List properties",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by ZIP code",
                        "name": "zip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by city",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by state code (e.g. FL)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by property type",
                        "name": "property_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum price in USD",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum price in USD",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum bedrooms",
                        "name": "min_beds",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum bathrooms",
                        "name": "min_baths",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Minimum home size in sqft",
                        "name": "min_sqft",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Maximum home size in sqft",
                        "name": "max_sqft",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "newest",
                            "price_asc",
                            "price_desc"
                        ],
                        "type": "string",
                        "default": "newest",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 24,
                        "description": "Page size (1-100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Opaque pagination cursor from a previous response's next_cursor",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.listResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/properties/{zpid}": {
            "get": {
                "description": "Full detail-screen payload for one listing, addressed by its Zillow property ID.\nmap_image_url is generated on first view; it may be null on the very first request for a listing and populated shortly after.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "properties"
                ],
                "summary": "Get property detail",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Zillow property ID",
                        "name": "zpid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.detailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.errorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.agentDTO": {
            "type": "object",
            "properties": {
                "brokerage": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "api.detailResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "agent": {
                    "$ref": "#/definitions/api.agentDTO"
                },
                "bathrooms": {
                    "type": "number"
                },
                "bedrooms": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "cooling": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "detail_url": {
                    "type": "string"
                },
                "garage": {
                    "type": "string"
                },
                "heating": {
                    "type": "string"
                },
                "hoa_fee_monthly": {
                    "type": "integer"
                },
                "home_size_sqft": {
                    "type": "integer"
                },
                "image_urls": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "latitude": {
                    "type": "number"
                },
                "listing_status": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                },
                "lot_size_acres": {
                    "type": "number"
                },
                "lot_size_sqft": {
                    "type": "integer"
                },
                "map_image_url": {
                    "type": "string"
                },
                "mls_number": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "property_type": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                },
                "year_built": {
                    "type": "integer"
                },
                "zip": {
                    "type": "string"
                },
                "zpid": {
                    "type": "string"
                }
            }
        },
        "api.errorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "api.listItem": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bathrooms": {
                    "type": "number"
                },
                "bedrooms": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "home_size_sqft": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "property_type": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                },
                "zpid": {
                    "type": "string"
                }
            }
        },
        "api.listResponse": {
            "type": "object",
            "properties": {
                "next_cursor": {
                    "type": "string"
                },
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.listItem"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}